CHI-405: fix Expat Liveability ROW_NUMBER cast + backfill analysis_cache nulls
Three of the four bugs listed in CHI-405 were already fixed in production;
the parallel lib/indicators/ copy had regressed and was the real risk.

- lib/indicators/expat-liveability.ts: cast ROW_NUMBER() to ::int in SQL AND
  wrap r.rn with Number() in the .find() comparison. Without both, postgres
  returns BIGINT (as BigInt/string in postgres.js) and strict equality silently
  fails, making score always null.
- supabase/functions/analyse-job/index.ts: add matching ::int cast so the
  edge function is defence-in-depth even without the JS Number() wrapper.

Also backfilled 7 old analysis_cache rows with null municipio/provincia via
PostGIS nearest-centroid join on municipios (ran directly against DB, no
migration needed). Corrected 1 row where a comunidad name had leaked into
the provincia column. Deleted INE 53053 from municipios — a Mancomunidad,
not a real municipio, that was polluting reverse-geocode nearest-neighbour
lookups near León.

Final state: 0 null municipio, 0 null provincia across analysis_cache;
8,339/8,339 municipios with full comunidad+provincia. 2 rows still have
null codigo_postal because postal_zones MVP covers only Málaga 29001–29017
— tracked as a postal_zones coverage gap, not a pipeline bug.

true_affordability is still null everywhere — that's CHI-400/401 scope.

Started M6_PROGRESS.md to track the sprint.

+++ supabase/functions/analyse-job/index.ts
@@ -949,6 +949,9 @@ async function calcExpatLiveability(
 ): Promise<IndicatorResult> {
   const alerts: Alert[] = []
 
+  // CHI-405: cast ROW_NUMBER() (BIGINT) to INT in SQL — without this, postgres
+  // returns a BigInt/string that fails strict equality in the Number(r.rn) === 1
+  // comparison below, and the function always returns score: null.
   const rows = await sql<{
     dist_m: number; nombre: string; iata_code: string; weekly_flights: number; rn: number
   }[]>`
@@ -956,7 +959,7 @@ async function calcExpatLiveability(
       SELECT
         ST_Distance(geom, ST_SetSRID(ST_MakePoint(${prop.lng}, ${prop.lat}), 4326)::GEOGRAPHY) AS dist_m,
         nombre, iata_code, weekly_flights,
-        ROW_NUMBER() OVER (ORDER BY geom <-> ST_SetSRID(ST_MakePoint(${prop.lng}, ${prop.lat}), 4326)::GEOGRAPHY) AS rn
+        ROW_NUMBER() OVER (ORDER BY geom <-> ST_SetSRID(ST_MakePoint(${prop.lng}, ${prop.lat}), 4326)::GEOGRAPHY)::int AS rn
       FROM airports
     )
     SELECT dist_m, nombre, iata_code, weekly_flights, rn
